requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.36
datasets>=2.16
accelerate>=0.26
ultralytics>=8.0
opencv-python>=4.8
pillow>=9.5
numpy>=1.24
pandas>=2.0
tqdm>=4.65
scikit-learn>=1.3
matplotlib>=3.7